+++ CHANGELOG.md
@@ -45,6 +45,7 @@ Changelog
    * moved all code to the `Puli\` top-level namespace
    * moved `ResourceRepository` and related classes to `Puli\Repository\`
    * renamed `getByTag()` to `findByTag()`
+   * added `getLongestCommonBasePath()` and `isBasePath()` to `Path`
 
 * 1.0.0-alpha3 (2014-02-22)